commit before branch change
