+++ src/components/Typography/styles.scss
@@ -24,8 +24,8 @@
 .t-h2 {
   font-family: $font-sans-serif;
   font-weight: 900;
-  font-size: 30px;
   color: $navy-black;
+  font-size: 30px;
   letter-spacing: 0;
   line-height: 40px;
   text-align: center;
